refactor: promote pkg/core to top-level core/
