+++ core/cache/graph/cache_test.go
@@ -2,8 +2,8 @@ package graph
 
 import (
 	"context"
-	"github.com/anaregdesign/lantern/pkg/core/cache"
-	"github.com/anaregdesign/lantern/pkg/core/graph"
+	"github.com/anaregdesign/lantern/core/cache"
+	"github.com/anaregdesign/lantern/core/graph"
 	"reflect"
 	"testing"
 	"time"
