+++ core/concurrent/pubsub/message.go
