+++ core/graph/view.go
