+++ core/nlp/dictionary_test.go
